Inspect project field keys (#8)
Jira allow users to create custom fields. The metadata about these
fields need to be fetched from the Jira web service. The fields are
stored in a nested structure, that we will need to parse later. For now,
we just want to parse the rest of the structure, in order to get to the
field data at all.
- Create `ProjectFieldKeys` for traversing the Jira metadata payload for
project
- In `JiraSystemConfigLoader`, create an `inspect` method and a few
supporting methods to output a report on which issue types use which
fields.

Example (notice that `Epic` has an extra field (`customfield_10001`)
that the other types do not.)
```
$ python main.py --action 'inspect'
                     - Bug       Sub-Task  Task      Epic      Story     
issuelinks           - 1         1         1         1         1         
summary              - 1         1         1         1         1         
assignee             - 1         1         1         1         1         
issuetype            - 1         1         1         1         1         
customfield_10001    -                               1                   
project              - 1         1         1         1         1         
parent               - 1         1         1         1         1         
description          - 1         1         1         1         1         
labels               - 1         1         1         1         1         
                     - Bug       Sub-Task  Task      Epic      Story     
```