Load generated models (#12)
We want to model the Jira data model locally, in order to mirror the
functionality of the website. With the newly created `plugins`
directory, users can now generate models and instantiate objects that
are based on the upstream data.
- Create a `module loader` (see `plugins/__init__.py`), allowing users
to dynamically load generated models: `import plugins`
- Create class inside the `plugins` module, because I'm unsure if I want
to to one or the other: `from plugins import Plugins`
- Create example plugin: `plugins/plugins_test.py`

We want to be able to automatically load all modules, and we want them
to be namespaced.

Either of these solutions work, I just haven't decided which way to go
yet.

```python
$ from plugins import Plugins
$ Plugins.all_plugins['plugins_test'].Schema(type='a', system='b')
type='a' system='b'

$ print(Plugins.plugins_test.Schema(type='a', system='b'))
type='a' system='b'

$ import plugins
$ plugins.plugins_test.Schema(type='a', system='b')
type='a' system='b'

$ from plugins.plugins_test import Schema
$ Schema(type='a', system='b')
type='a' system='b'
```