Default value overwrite None (#39)
When the key exists in a dict and the key refers to `None`, then the
value returned from `get(key, default)` will be `None`, even if a
`default` is provided. To get around this, we introduce an explicit
overwrite.

```diff
      def get(self, key: str, default: Any = None) -> dict | None:
-         return self.data.get(key, default)
+         return self.data.get(key, default) or default
```

This means that we can rely on the default value of `{}` as below,
without worrying about `assignee_name` being `None`.

```python
        assignee = self.issue.get_field("assignee", {})
        assignee_name = assignee.get("displayName")
```