Ensure optional allowed_types is always set to a value (#58)
The `JiraIssues.allowed_types` class variable is optional:

```python
class JiraIssues:
    _allowed_types: list[str] | None = None
```

We wan't it to stay nullable, in order to check for upstream issues when
setting it. But we never actually want it to be `None`.

This PR forces a `ValueError` if the variable is about to be set to
`None`. This allows us to remove fallbacks like this:
```python
assert 'Story' in (allowed_types or [])
```