Get field raise for non existent key (#65)
`JiraIssue.fields` is a `dict` with data coming from upstream. Any field
that is not set upstream is, nonetheless, available in this payload, but
its value will be `null`. Therefore, in cases where a field is not
listed in the payload as a nulled value, we know that it is not
available upstream.

In order to reflect this fact, in this PR we change how we handle this
`None`. Where, previously, we would allow the getter to fall back to the
default if a value was not set, going forward, if a key is not set in
the payload, we will be raising a `ValueError`.

Raising early will ensure that we won't be attempting to model or update
any fields that don't exist upstream.