Issue.get_field to not raise ValueError (#78)
Instead of raising `ValueError` when a field does not exist in the
(cached) upstream issue data, we instead return `None` or the `default`
value. This means we will need to do `None`-checks elsewhere to avoid
preparing bad updates. However, this is what the `createmeta` and
`editmeta` endpoints are for, so we should not rely on the issue data to
decide which fields exist and which ones don't.