Fix createmeta cache write location and typing (#91)
In `jira_system_config_loader.py`, the `get_createmeta` method was
writing to the wrong file.
After getting the `createmeta` from upstream, it was written to disk
using `write_issuetypes_to_system_cache`.
```diff
-         issuetypes = self.client.get_createmeta(issuetype_id)
+         createmeta = self.client.get_createmeta(issuetype_id)
  ...
-         self.cache.write_issuetypes_to_system_cache(issuetypes)
-         return issuetypes
+         self.cache.write_createmeta(issuetype_name, createmeta)
+         return createmeta
```
This means that `createmeta` was written correctly, and the `issuetypes`
would by corrupted.

Since the schemas of the `issuetypes` object and `createmeta` differ,
this caused a number of typing issues across the repo as well. I'm
fixing that while I'm at it.