Fix: Let draft template be loaded as a Python file (#139)
Since we are aiming for being able to run the CLI anywhere, the draft
template needs to make sure the template can be read. Ideally, the
content should be configurable client side, but for now we'll just
import it as a Python file, so the content is at least loading.